+++ knudu/appJs/Login.js
@@ -1,0 +1,99 @@
+import React, {Component} from 'react';
+import {Alert, Button, TextInput, Platform, StyleSheet, Text, View} from 'react-native';
+
+const instructions = Platform.select({
+  ios: 'Press Cmd+R to reload,\n' + 'Cmd+D or shake for dev menu',
+  android:
+    'Double tap R on your keyboard to reload,\n' +
+    'Shake or press menu button for dev menu',
+});
+
+export default class Login extends Component{
+  constructor(props) {
+    super(props);
+    
+    this.state = {
+      username: '',
+      password: '',
+    };
+  }
+  
+  onLogin() {
+    const { username, password } = this.state;
+    Alert.alert('Credentials', `${username} + ${password}`);
+  }
+
+  render() {
+    return (
+      <View style={styles.container}>
+        <Text style={styles.ueberschrift}>
+          Login
+        </Text>
+        <TextInput
+          value={this.state.username}
+          onChangeText={(username) => this.setState({ username })}
+          placeholder={'Username'}
+          style={styles.input}
+        />
+        <TextInput
+          value={this.state.password}
+          onChangeText={(password) => this.setState({ password })}
+          placeholder={'Password'}
+          secureTextEntry={true}
+          style={styles.input}
+        />
+        <View  style={styles.buttonContainer}>
+          <Button
+            title={'Login'}
+            style={styles.button}
+            onPress={this.onLogin.bind(this)}
+          />
+
+          <Button
+            title={'Registrieren'}
+            titleStyle={styles.buttonTitle}
+            style={styles.button}
+            onPress={this.onLogin.bind(this)}
+          />
+        </View>
+      </View>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#bdc3c7',
+  },
+  ueberschrift:{
+    fontSize: 30,
+    marginBottom: 40,
+  },
+  instructions: {
+    textAlign: 'center',
+    color: '#333333',
+    marginBottom: 5,
+  },
+  input: {
+    width: 200,
+    height: 44,
+    padding: 10,
+    borderWidth: 1,
+    borderColor: 'black',
+    marginBottom: 10,
+  },
+  buttonContainer:{
+    justifyContent: 'center',
+    alignItems: 'center',
+    flexDirection: 'row',
+  },
+  button: {
+    padding: 10,
+    borderWidth: 2,
+    margin: 20,
+  },
+});
